move client file to client directory
