--- App.js
@@ -1,10 +1,0 @@
-import React from "react";
-import MainContainer from "./components/MainContainer";
-
-
-function App() {
-    return <MainContainer />;
-
-}
-
-export default App;
