+++ client/src/components/SearchArea.js
